fix: Avoid conflicts between global support args
The `--print-comment-definition` option is available on every check
command to output an Icinga-compatible command definition. The code did
check for that argument before adding all other arguments, to avoid
conflicts with later added required arguments using `#parse_known_args`.

That in turn conflicted with `--help`, which was handled internally by
`argparse` when parsing the first time, so in `#parse_known_args`. At
that point, nothing except `--print-comment-definition` was added yet
and the printed help was empty.

This CL refactors `--print-comment-definition` to us an `argparse`
action, similar to the included `--help` action, that internally prints
the result and aborts parsing. This way, we do not need to parse upfront
nor will conflict with required arguments.